requirements.txt
flask
numpy
torch
torchvision
Pillow
gunicorn
colormath
colorthief
colour
opencv-contrib-python
pandas
scipy
